TP-760: Move old classes to junit4 package
* JUnit4 classes are moved to subpackage junit4
* Performed in separate commit to preserve git history
